+++ metrox-android/src/main/kotlin/dev/zacsweers/metrox/android/ActivityKey.kt
@@ -1,12 +1,12 @@
 // Copyright (C) 2025 Zac Sweers
 // SPDX-License-Identifier: Apache-2.0
-package dev.zacsweers.metro.sample.androidviewmodel.components
+package dev.zacsweers.metrox.android
 
 import android.app.Activity
 import dev.zacsweers.metro.MapKey
 import kotlin.reflect.KClass
 
-/** A [MapKey] annotation for binding Activities in a multibinding map. */
+/** A [MapKey] annotation for binding an Activity in a multibinding map. */
 @MapKey
 @Target(AnnotationTarget.CLASS)
-annotation class ActivityKey(val value: KClass<out Activity>)
+public annotation class ActivityKey(val value: KClass<out Activity>)