+++ samples/android-app/src/main/kotlin/dev/zacsweers/metro/sample/android/AppGraph.kt
@@ -2,7 +2,6 @@
 // SPDX-License-Identifier: Apache-2.0
 package dev.zacsweers.metro.sample.android
 
-import android.app.Activity
 import android.app.Application
 import android.content.Context
 import androidx.work.ListenableWorker
@@ -12,18 +11,13 @@ import dev.zacsweers.metro.DependencyGraph
 import dev.zacsweers.metro.Multibinds
 import dev.zacsweers.metro.Provider
 import dev.zacsweers.metro.Provides
+import dev.zacsweers.metrox.android.MetroAppComponentProviders
 import kotlin.reflect.KClass
 
 @DependencyGraph(AppScope::class)
-interface AppGraph {
+interface AppGraph : MetroAppComponentProviders {
 
   @Provides fun provideApplicationContext(application: Application): Context = application
-
-  /**
-   * A multibinding map of activity classes to their providers accessible for
-   * [MetroAppComponentFactory].
-   */
-  @Multibinds val activityProviders: Map<KClass<out Activity>, Provider<Activity>>
 
   val workManager: WorkManager