+++ samples/compose-navigation-app/src/main/kotlin/dev/zacsweers/metro/sample/androidviewmodel/MainActivity.kt
@@ -23,8 +23,8 @@ import dev.zacsweers.metro.AppScope
 import dev.zacsweers.metro.ContributesIntoMap
 import dev.zacsweers.metro.Inject
 import dev.zacsweers.metro.binding
-import dev.zacsweers.metro.sample.androidviewmodel.components.ActivityKey
 import dev.zacsweers.metro.sample.androidviewmodel.viewmodel.metroViewModel
+import dev.zacsweers.metrox.android.ActivityKey
 import kotlinx.serialization.Serializable
 
 @ContributesIntoMap(AppScope::class, binding<Activity>())