+++ samples/compose-navigation-app/src/main/kotlin/dev/zacsweers/metro/sample/androidviewmodel/components/AppGraph.kt
@@ -4,25 +4,16 @@
 
 package dev.zacsweers.metro.sample.androidviewmodel.components
 
-import android.app.Activity
 import dev.zacsweers.metro.AppScope
 import dev.zacsweers.metro.DependencyGraph
-import dev.zacsweers.metro.Multibinds
-import dev.zacsweers.metro.Provider
 import dev.zacsweers.metro.Provides
 import dev.zacsweers.metro.SingleIn
 import dev.zacsweers.metro.sample.androidviewmodel.viewmodel.ViewModelGraph
+import dev.zacsweers.metrox.android.MetroAppComponentProviders
 import kotlin.concurrent.atomics.AtomicInt
 import kotlin.concurrent.atomics.ExperimentalAtomicApi
-import kotlin.reflect.KClass
 
 @DependencyGraph(AppScope::class)
-interface AppGraph : ViewModelGraph.Factory {
-  /**
-   * A multibinding map of activity classes to their providers accessible for
-   * [MetroAppComponentFactory].
-   */
-  @Multibinds val activityProviders: Map<KClass<out Activity>, Provider<Activity>>
-
+interface AppGraph : ViewModelGraph.Factory, MetroAppComponentProviders {
   @Provides @SingleIn(AppScope::class) fun provideViewModelCounter(): AtomicInt = AtomicInt(0)
 }